% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers.R
\name{add_dcp_start}
\alias{add_dcp_start}
\alias{add_ig_start}
\alias{add_powell_ig_tiers}
\alias{add_mead_ig_tiers}
\title{Add IG and DCP lines and annotation}
\usage{
add_dcp_start(gg)

add_ig_start(gg)

add_powell_ig_tiers(gg)

add_mead_ig_tiers(gg)
}
\arguments{
\item{gg}{Object inheriting from \code{gg}.}
}
\value{
\code{gg} object
}
\description{
For Lakes Powell and Mead, \code{add_powell_ig_tiers()} and \code{add_mead_ig_tiers()}
add in the key elevations that distinguish the different operating tiers as
specified by the IGs. Additionally, \code{add_dcp_start()} and \code{add_ig_start()}
add vertical markers for the years that the IGs and DCPs started. In addition
to the lines, annotation text describes each line.
}
