% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{compare_crit_stats}
\alias{compare_crit_stats}
\title{Probability plot for single variable and multiple scenarios}
\usage{
compare_crit_stats(
  zz,
  yrs,
  variable,
  annText,
  plotTitle,
  legendTitle = "",
  legLoc = "right",
  nC = 1,
  annSize = 3,
  legendWrap = NULL,
  plot_colors = NULL
)
}
\arguments{
\item{zz}{Data frame. Must have Year, Variable, AggName, and Value columns.}

\item{yrs}{Years to show in plot. \code{zz} is filtered to only contain these
years.}

\item{variable}{Variable to plot. \code{zz} is filtered to only contain this
variable.}

\item{annText}{Caption. Used in \code{annotation()} for a y value of 0.95.
Is this more a subtitle? Test it out?}

\item{plotTitle}{Title. Used in \code{labs(title = plotTitle)}.}

\item{legendTitle}{Color legend title.}

\item{legLoc}{Color legend location. Defaults to right.}

\item{nC}{Number of columns in the color legend.}

\item{annSize}{Text size for \code{annText}.}

\item{legendWrap}{Maximum number of character per line in color legend.
StartMonth will be wrapped based on this value, if it is not \code{NULL}.}

\item{plot_colors}{Optional named vectors. If specified, names must match
unique StartMonth values and can be used to set specific colors for each
scenario. If not used, default ggplot2 colors are used.}
}
\value{
\code{gg} object.
}
\description{
\code{compare_crit_stats()} compares the probability that a single event
(variable) occurs/does not occur for multiple scenarios.
}
