% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_results_package.R
\name{create_results_package}
\alias{create_results_package}
\title{Create CRSS results package}
\usage{
create_results_package(file)
}
\arguments{
\item{file}{A yaml file containing user input instructions for the CRSS
results package.}
}
\description{
\code{create_results_package()} creates a "package" of CRSS results. The package
consists of the user specified figures and comparison(s) accross one or
more scenarios. All user input provided by the yaml \code{file}. See
\url{https://github.com/BoulderCodeHub/Process-CRSS-Res/wiki/yaml-specification}
for documentation on specifying the yaml.
}
