% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_boxplots.R
\name{csd_bxp}
\alias{csd_bxp}
\title{Annual boxplot for one variable and multiple scenarios}
\usage{
csd_bxp(zz, state, yrs, plot_colors)
}
\arguments{
\item{zz}{Data frame. Must have Variable, Year, Value, and Agg columns.}

\item{state}{One of: c("CO", "NM", "UT", "WY","AZ", "NV", "CA", "MX").}

\item{yrs}{The years to show in plot. \code{zz} is filtered to only contain
these years.}

\item{plot_colors}{Named vector. Names must match unique Agg values. Sets
plots to use specific colors for each scenario.}
}
\value{
\code{gg} object.
}
\description{
\code{csd_bxp()} creates an annual boxplot for one variable and multiple
scenarios. It uses all scenarios found in \code{zz}. It could be more generic,
except that it relies on \code{csd_state_info()} that maps state names, to their
computed state depletions slot, and includes their apportionment.
}
