% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spec.R
\name{csd_spec}
\alias{csd_spec}
\title{csd_spec is a class for specifying the annual computed state depletions plots}
\usage{
csd_spec(ui, pg)
}
\description{
csd_spec is a class for specifying the annual computed state depletions plots
}
