% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_pe}
\alias{ex_pe}
\title{Example data}
\format{
Data frame with 21,146 rows and 7 variables:
\describe{
\item{Year}{year}
\item{Month}{month, full name}
\item{TraceNumber}{trace number, 1-31 are from CRSS, 0 are initial
conditions}
\item{Scenario}{RiverSMART scenario folder, 2 are included}
\item{Variable}{variable name, 11 are included}
\item{Value}{variable value, in feet for \code{powell_dec_pe} and \code{mead_dec_pe}
and binary occurrence for remaining variables.}
\item{ScenarioGroup}{scenario group. Used to group multiple "Scenario"
folders together for processing/analysis. 2 are included}
}
}
\usage{
ex_pe
}
\description{
\code{ex_pe} is a data frame (tibble) that includes example data based on Lakes
Powell and Mead pool elevation for multiple scenarios and metrics. Results
are from April 2020 CRSS using the stress test hydrology, and were generated
using \code{\link[RWDataPlyr:rw_scen_aggregate]{RWDataPlyr::rw_scen_aggregate()}}.
}
\keyword{datasets}
