% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_yaml_input.R
\name{expand_plot_group}
\alias{expand_plot_group}
\title{Takes one plot_group, and fully expands it, inheriting defaults from defaults
plot_group (\code{pg}) will have scenarios, scen_names, plot_colors, years,
caption entries after it is expanded}
\usage{
expand_plot_group(pg, defaults)
}
\description{
Takes one plot_group, and fully expands it, inheriting defaults from defaults
plot_group (\code{pg}) will have scenarios, scen_names, plot_colors, years,
caption entries after it is expanded
}
