% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_get_legend.R
\name{gg_get_legend}
\alias{gg_get_legend}
\title{Get the legend from ggplot}
\usage{
gg_get_legend(gg)
}
\arguments{
\item{gg}{object inheriting from \code{gg}.}
}
\value{
gtable, grob
}
\description{
\code{gg_get_legend()} gets the legend from a ggplot object and returns as a
gtable grob.
}
