% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spec.R
\name{inherit_from_pg}
\alias{inherit_from_pg}
\title{sets the specified \code{att} in \code{x}. Checks to see if it exists in \code{ui}. If it
does not, then defaults to \code{att} in \code{pg}.}
\usage{
inherit_from_pg(x, att, ui, pg, plot_type)
}
\description{
sets the specified \code{att} in \code{x}. Checks to see if it exists in \code{ui}. If it
does not, then defaults to \code{att} in \code{pg}.
}
