% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_yaml_input.R
\name{parse_yaml_input}
\alias{parse_yaml_input}
\title{Parse yaml input

\code{parse_yaml_input()} parses the yaml input into the data structure required
by \code{\link[=process_everything]{process_everything()}}.}
\usage{
parse_yaml_input(file)
}
\arguments{
\item{file}{Path to the yml file.}
}
\description{
Parse yaml input

\code{parse_yaml_input()} parses the yaml input into the data structure required
by \code{\link[=process_everything]{process_everything()}}.
}
