% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group.R
\name{plot_group}
\alias{plot_group}
\title{creates a \code{plot_group} object from a list \code{x}. \code{x} must have certain entries
otherwise the conversion will fail}
\usage{
plot_group(x, defaults)
}
\description{
Required elements in x: scenarios
Will fill in: scen_names, plot_colors, years, caption
Will create std_comparison, csd_ann, heat, cloud plot types and plots entry
}
