% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCloudFig.R
\name{plotCloudFigs}
\alias{plotCloudFigs}
\title{Plot range of conditions as cloud}
\usage{
plotCloudFigs(zz, yrs, var, myTitle, ui, pg_i)
}
\arguments{
\item{zz}{Data frame. Must have StartMonth, Year, Variable, and Value
columns.}

\item{yrs}{Years to show in the plot. If any exist before the min in \code{zz},
then historical values are obtained and added to the figure.}

\item{var}{Should be either "powell_dec_pe" or "mead_dec_pe". \code{zz} is
filtered to only include those data.}

\item{myTitle}{Plot title.}

\item{ui}{List. Currently requires entries that have scenarios, scenario
names, color label, and legend wrapping numbers.}

\item{pg_i}{Integer. Index into \code{ui} used for getting the scenarios and
scenario names.}
}
\value{
\code{ggGrob}
}
\description{
\code{plotCloudFigs()} plots the 10th and 90th percentiles as a cloud and the
median as a line. It will also add in historical data, and is hard coded
to only work for Powell and Mead elevation. Function adds in reclamation
object.
}
