% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotCritStats}
\alias{plotCritStats}
\title{Plot the "critical stats" figure}
\usage{
plotCritStats(zz, yrs, annText, legendTitle = "", legLoc = "bottom", nC = 4)
}
\arguments{
\item{zz}{Data frame. Requires Year, Variable, Value columns. All other
columns are dropped when grouped by Year and Variable and then summarised.
The Variable columns is expected to have a subset of variables in
\code{csVarNames()}. See details.}

\item{yrs}{Vector of years to show in plot. \code{zz} is filtered to only include
these years.}

\item{annText}{Used in labs(caption=annText)}

\item{legendTitle}{How to title the color legend.}

\item{legLoc}{The location of the color legend.}

\item{nC}{The number of columns to use in color legend.}

\item{annSize}{Unused. Was the size of the annotation.}
}
\value{
\code{gg} object.
}
\description{
\code{plotCritStats()} creates a figure with multiple variables and one scenario
where the variables are the probability a condition will/will not occur.
}
\details{
Expected Variables are: lbShortage, lbSurplus, mead_min_lt_1000,
mead_min_lt_1020, mead_min_lt_1025, powell_wy_min_lt_3490,
powell_wy_min_lt_3525, mead_dec_lt_1000, mead_dec_lt_1020,
powell_dec_lt_3525, mead_dec_lt_1025.
}
