% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotEOCYElev}
\alias{plotEOCYElev}
\title{Plot 10/50/90 percentiles}
\usage{
plotEOCYElev(
  zz,
  yrs,
  var,
  myTitle,
  legendTitle,
  legendWrap = NULL,
  plot_colors = NULL
)
}
\arguments{
\item{zz}{Data frame. Must have Year, Variable, StartMonth, and Value
columns.}

\item{yrs}{Years to show in plot. \code{zz} is filtered to only contain these
years.}

\item{var}{Variable to plot. \code{zz} is filtered to only contain this variable.}

\item{myTitle}{Title. Used in \code{labs(title = myTitle)}.}

\item{legendTitle}{Color legend title.}

\item{legendWrap}{Maximum number of character per line in color legend.
StartMonth will be wrapped based on this value, if it is not \code{NULL}.}

\item{plot_colors}{Optional named vectors. If specified, names must match
unique StartMonth values and can be used to set specific colors for each
scenario. If not used, default ggplot2 colors are used.}
}
\value{
\code{gg} object.
}
\description{
\code{plotEOCYElev()} computes and then plots the 10, 50, and 90 percentiles for
the specified variable (\code{var}). Plots for different scenarios, denoted by
\code{StartMonth} column; all unique values in the StartMonth column are included
in the plot.
}
