% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFirstYearShortCond.R
\name{plotFirstYearShortCond}
\alias{plotFirstYearShortCond}
\title{Scatter plot of elevation vs. trace number}
\usage{
plotFirstYearShortCond(
  model,
  iFile,
  scenario,
  filterOn = "shortage",
  dataYear,
  colorVar = "WYRelease"
)
}
\arguments{
\item{model}{Either "MTOM" or "CRSS"}

\item{iFile}{Folder path or file that contains the data. If \code{model} is
CRSS, then the folder should have "crssShortCond_ann.feather" and
"crssShortCond_mon.feather" files. If \code{model} is "MTOM", then this should
be a csv file that contains the necessary data.}

\item{scenario}{If \code{model} is CRSS, then this is the scenario that will be
used.}

\item{filterOn}{Valid choices are 'shortage' or 'pe'. If 'pe', then will keep
all traces that are <= 1077.}

\item{dataYear}{Year to plot. Data are filtered to this year.}

\item{colorVar}{Has to be "WYRelease" if \code{model} is MTOM. Otherwise, can be
"WYRelease" or any other variable in the feather files. Typically, the other
value is "mwdIcs".}
}
\value{
\code{gg} object.
}
\description{
\code{plotFirstYearShortCond()} plots elevation vs. trace number for a single
year. It colors the points based on Powell release for the following WY and
the shape is based on either Powell October-December release or MWD's ICS
creation/delivery volume.
}
