% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotShortageSurplus}
\alias{plotShortageSurplus}
\title{Plot chances of LB shortage and surplus}
\usage{
plotShortageSurplus(
  zz,
  yrs,
  monthRun,
  legendTitle = "",
  nC = 2,
  legLoc = "bottom"
)
}
\arguments{
\item{zz}{Data frame. Must have Year, Variable, Value columns. Any others
will be collapsed together when summarizing. Variable must have
"lbShortage" and "lbSurplus" variables.}

\item{yrs}{The years that are shown in the figure.}

\item{monthRun}{String that will be added to the title.}

\item{legendTitle}{Title of the color legend.}

\item{nC}{The number of columns in the color legend.}

\item{legLoc}{The location of the color legend.}
}
\value{
\code{gg} object.
}
\description{
\code{plotShortageSurplus()} creates a line plot of the chances of shortage and
surplus for the specified years.
}
