% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotShortStackedBar}
\alias{plotShortStackedBar}
\title{Plot a stacked bar plot of shortage tiers}
\usage{
plotShortStackedBar(zz, yrs, annText)
}
\arguments{
\item{zz}{Data frame. Must have Year, Variable, and Value. All other columns
are collapsed together when summarizing. Variables must include
lbShortageStep1, lbShortageStep2, and lbShortageStep3.}

\item{yrs}{Years to show in plot. \code{zz} is filtered to only contain these
years.}

\item{annText}{Used in \code{labs(caption = annText)}.}
}
\value{
\code{gg} object.
}
\description{
\code{plotShortStackedBar()} plots a stacked barplot for the chances of shortage
for one scenario.
}
