% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pe_rwa}
\alias{pe_rwa}
\alias{short_cond_rwa}
\title{RiverWare data aggregator objects used by crssplot}
\format{
An object of class \code{rwd_agg} (inherits from \code{data.frame}) with 11 rows and 7 columns.

An object of class \code{rwd_agg} (inherits from \code{data.frame}) with 6 rows and 7 columns.
}
\usage{
pe_rwa

short_cond_rwa
}
\description{
\code{pe_rwa} is an \link[RWDataPlyr:rwd_agg]{RWDataPlyr::rwd_agg} object containing the common stats that
will be computed for Powell and Mead.

\code{short_cond_rwa} is an \link[RWDataPlyr:rwd_agg]{RWDataPlyr::rwd_agg} object containing containing the
annual stats that are necessary for the "conditions leading to shortage"
figure.
}
\details{
\code{pe_rwa} includes:
\itemize{
\item Powell and Mead EOCY elevation
\item Mead December elevation <= 1,000', 1,020', and 1,025'
\item Mead minimum yearly elevation <=  1,000', 1,020', and 1,025'
\item Powell December elevation <= 3,525'
\item Powell minimum water year elevation <= 3,490' and 3,525'
}

\code{short_cond_rwa} includes:
\itemize{
\item Mead EOCY elevation
\item Powell WY release
\item Annual total natural flow between Powell and Mead
\item Annual total natural flow below Mead
\item MWD annual creation and delivery (two different variables)
}
}
\keyword{datasets}
