% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scens_plot_probs.R, R/scens_plot_range.R
\name{scens_plot_probs}
\alias{scens_plot_probs}
\alias{scens_plot_range}
\title{Plot scenario comparison figures}
\usage{
scens_plot_probs(
  df,
  vars,
  years = NULL,
  scenarios = NULL,
  plot_colors = NULL,
  scen_labels = NULL,
  ...
)

scens_plot_range(
  df,
  vars,
  years = NULL,
  scenarios = NULL,
  plot_colors = NULL,
  scen_labels = NULL,
  ...
)
}
\arguments{
\item{df}{Data frame. Must have "Year", "Variable", "ScenarioGroup", and
"Value" columns.}

\item{vars}{Character vector specifying the variable(s) to use.}

\item{years}{Numeric vector specifying the years to show. If \code{NULL}, use all
years in \code{df}.}

\item{scenarios}{Character vector specifying the scenarios to use (found in
\code{df$ScenarioGroup}). If \code{NULL}, use all scenarios in \code{df}.}

\item{plot_colors}{Named character vector to set custom plot colors. Names
should match scenarios found in \code{df$ScenarioGroup}.}

\item{scen_labels}{Named character vector to set custom legend labels for the
scenarios. Used to show legend labels that are different from values found
in \code{df$ScenarioGroup}.}

\item{...}{Parameters passed to other functions. See details.}
}
\value{
\code{gg} object.
}
\description{
\code{scens_plot_probs()} plots probability plots, i.e., the chance of a variable
occurring. Different \code{scenarios} are shown as different colors, and if there
are different variables (\code{vars}) they are shown as different facets.

\code{scen_plot_range()} plots the range of results for multiple scenarios. The
range is shown as lines for 10th, 50th, and 90th percentiles. Typically
this is done for only one variable, but multiple variables can be provided
and will be shown as separate facets.
}
\details{
\code{...} is used to pass additional options to ggplot functions. The following
are used:
\itemize{
\item "title", and "caption" are passed to \code{\link[ggplot2:labs]{ggplot2::labs()}}. "color_label" and
"y_lab" are also passed using \code{color} and \code{y} parameters, respectively.
\item "legend_wrap" is used to modify labels in legends etc. by calling
\code{\link[stringr:str_wrap]{stringr::str_wrap()}} on the appropriate variable, with
\code{width = legend_wrap}.
\item "facet_scales", "facet_nrow", and "facet_ncol" are all passed to
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.
}

\code{scens_plot_probs()} and \code{scens_plot_range()} will use the following
additional options: "y", "title", "color_label", "legend_wrap",
"facet_scales", "facet_nrow", and "facet_ncol".

The legend order can be modified by converting "ScenarioGroup" column to a
factor before calling \verb{scens_plot_*()}, with the levels specifying the
order the scenarios will show up in the legend.
}
\examples{
# quick simple plot of one variable
scens_plot_range(ex_pe, "mead_dec_pe"))

# now add ylabel 
scens_plot_range(ex_pe, "mead_dec_pe", y_lab = "feet")

# show two variables
scens_plot_range(
  ex_pe, 
  c("powell_dec_pe", "mead_dec_pe"), 
  facet_scales = "free_y"
)

# subset scenarios
scens_plot_range(ex_pe, "mead_dec_pe", scenarios = "April ST CT"))

# custom colors and scenario labels + add title and caption
pc <- c("April ST CT" = "red", "April ST 2007 UCRC" = "black")
sl <- c("April ST CT" = "s1", "April ST 2007 UCRC" = "s2")
scens_plot_range(ex_pe, 
  "powell_dec_pe", 
  plot_colors = pc, 
  scen_labels = sl,
  title = "PE", 
  caption = "this is a caption")
)

# change to two rows in stead of two columns for showing two variables
scens_plot_range(
  ex_pe, 
  c("powell_dec_pe", "mead_dec_pe"), 
  facet_scales = "free_y",
  facet_nrow = 2)
)

}
