% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{singleYearPEScatter}
\alias{singleYearPEScatter}
\title{Scatter plot for one year of Mead elevation}
\usage{
singleYearPEScatter(zz, yr, var, myTitle, caption = NULL, addThreshStats)
}
\arguments{
\item{zz}{Data frame. Must have Year, Variable, Value, and TraceNumber
columns.}

\item{yr}{Year to create plot for.}

\item{var}{The variable to plot. This is not necessary, as the other
components are so specific that they can really only apply to Mead
elevation.}

\item{myTitle}{Title of plot.}

\item{caption}{Caption (\code{labs(caption = caption)}).}

\item{addThreshStats}{Boolean. If \code{TRUE}, annotates figure with the number
of traces that fall into each of the four elevation bins.}

\item{return}{\code{gg} object.}
}
\description{
\code{singleYearPEScatter()} plots elevation vs. trace number for one year. It
specifically colors points based on pre-specified elevation thresholds:
below 1,075', 1,075'-1,076', 1,076'-1,077', and above 1,077'.
}
