% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spec.R
\name{std_comp_spec}
\alias{std_comp_spec}
\title{std_comp_spec is a class for specifying the standard comparison plots}
\usage{
std_comp_spec(ui, pg)
}
\description{
std_comp_spec is a class for specifying the standard comparison plots
}
