% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_condition_heat_map.R
\name{system_conditions_heat_map}
\alias{system_conditions_heat_map}
\title{Create a heat map of conditions}
\usage{
system_conditions_heat_map(zz, n_yrs, tier_names, my_title, y_title, heat_ui)
}
\arguments{
\item{zz}{Data frame with Year, Variable, Value, and Agg columns.}

\item{n_yrs}{Unnecessary parameter. Could be computed as
length(unique(zz$Year))}

\item{tier_names}{Unnecessary parameter. Used to determine number of tiers,
but it can also be computed as length(unique(zz$Variable)).}

\item{my_title}{Used in labs(title)}

\item{y_title}{Used in labs(y)}

\item{heat_ui}{Only \verb{$caption} entry is accessed directly.}
}
\description{
\code{system_condition_heat_map()} creates a heat map of conditions
(\code{zz$Variable}) for all years (\code{zz$Year}) and scenarios (\code{zz$Agg}) in \code{zz}.
}
