% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_crss.R
\name{theme_crss}
\alias{theme_crss}
\alias{theme_cloud}
\title{Customized ggplot2 themes}
\usage{
theme_crss()

theme_cloud()
}
\description{
\code{theme_crss()} and \code{theme_cloud()} are custom themes used in typical CRSS
related plots.
}
